+++ src/app/core/services/api/fake-backend/broker-fake-impl/catalog-fake-impl.ts
@@ -112,6 +112,20 @@ export const getCatalogPage = (query: CatalogPageQuery): CatalogPageResult => {
             {id: '', title: ''},
           ],
         },
+        {
+          id: 'connectorEndpoint',
+          title: 'Connector',
+          values: [
+            {
+              id: 'https://example-connector/api/dsp',
+              title: 'https://example-connector/api/dsp',
+            },
+            {
+              id: 'https://example-connector2/api/dsp',
+              title: 'https://example-connector2/api/dsp',
+            },
+          ],
+        },
       ],
     },
     paginationMetadata: {